third_party/WebKit/LayoutTests/editing/selection/move-by-line-004.html

<script src="../../fast/js/resources/js-test-pre.js"></script>
<script src="./resources/js-test-selection-shared.js"></script>

<div id="test" contenteditable="true">
    <div contenteditable="false">
        <span id="pie" contenteditable="true">apple pie</span>
    </div>
    <div contenteditable="false">
        pineapple
    </div>
    <div contenteditable="false">
        <span id="juice" contenteditable="true">apple juice</span>
    </div>
</div>

<script>
description('Test that moving the caret with line granularity within the highest editable root doesn\'t work across non-editable content.');

testCaretMotion({
    'direction': 'forward',
    'granularity': 'line',
    'origin': 'pie',
    'originOffset': 3,
    'target': 'pie',
    'targetOffset': -1,
});

testCaretMotion({
    'direction': 'backward',
    'granularity': 'line',
    'origin': 'juice',
    'originOffset': 3,
    'target': 'juice',
    'targetOffset': 0,
});

if (window.testRunner)
    $('test').outerHTML = '';
</script>
<script src="../../fast/js/resources/js-test-post.js"></script>
