third_party/WebKit/LayoutTests/editing/selection/move-by-line-004-expected.txt

Test that moving the caret with line granularity within the highest editable root doesn't work across non-editable content.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Move forward line x 1
PASS selection.anchorNode == '[object Text]' is true
PASS selection.anchorOffset is 9
Move backward line x 1
PASS selection.anchorNode == '[object Text]' is true
PASS selection.anchorOffset is 0
PASS successfullyParsed is true

TEST COMPLETE
